chore: refactor to add logic for additonal option
