+++ app/lib/theme/widgets/SelectWidget.tsx
@@ -1,13 +1,40 @@
 import { WidgetProps } from '@rjsf/core';
+import { Dropdown } from '@button-inc/bcgov-theme';
 import styled from 'styled-components';
-import StyledSelect from '../components/StyledDropdown';
 import { Label } from '../../../components/Form';
 
+const StyledSelect = styled(Dropdown)`
+  .pg-select-wrapper {
+    margin: 12px 0;
+    width: ${(props) => props.theme.width.inputWidthSmall};
+  }
+  .pg-select-wrapper:after {
+    margin: 0.3em 0;
+  }
+  & select {
+    margin: 0.25em 0;
+  }
+
+  select:disabled {
+    background: inherit;
+    border: 1px solid rgba(96, 96, 96, 0.3);
+    opacity: 0;
+  }
+
+  & div:first-child {
+    background: ${(props) => props.disabled && 'rgba(196, 196, 196, 0.3)'};
+    border: ${(props) => props.disabled && ' 1px solid rgba(96, 96, 96, 0.3)'};
+  }
+`;
 const StyledDiv = styled('div')`
   margin-bottom: 32px;
 `;
 
-const SelectWidget: React.FC<WidgetProps> = ({
+interface SelectWidgetProps extends WidgetProps {
+  customOption?: React.ReactNode;
+}
+
+const SelectWidget: React.FC<SelectWidgetProps> = ({
   id,
   disabled,
   onChange,
@@ -17,6 +44,7 @@ const SelectWidget: React.FC<WidgetProps> = ({
   placeholder,
   schema,
   uiSchema,
+  customOption,
 }) => {
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
@@ -46,6 +74,7 @@ const SelectWidget: React.FC<WidgetProps> = ({
               {opt}
             </option>
           ))}
+        {customOption ?? customOption}
         {description && <Label>{description}</Label>}
       </StyledSelect>
     </StyledDiv>
